Añadida cinemática directa y otros cambios menores

+++ tripteron_firmware/inc/drivers/encoder/encoder_interface.h
@@ -8,7 +8,7 @@ class IEncoder {
 public:
     virtual ~IEncoder() = default;
     virtual bool SetOffset() = 0;
-    virtual bool ReadAngle(double& angle) = 0;
+    virtual bool ReadAngle(float& angle) = 0;
 };
 
 std::shared_ptr<IEncoder> MakeIEncoder(std::shared_ptr<II2CWrapper> i2cWrapper);
